Move parse_OMCError
